BOAC-5777, break cohort_controller into three controllers (readability, etc.)
