+++ src/components/cohort/CohortPageHeader.vue
@@ -203,7 +203,8 @@ import ExportListModal from '@/components/util/ExportListModal.vue'
 import FerpaReminderModal from '@/components/util/FerpaReminderModal.vue'
 import RenameCohort from '@/components/cohort/RenameCohort.vue'
 import {alertScreenReader, pluralize, putFocusNextTick} from '@/lib/utils'
-import {deleteCohort, downloadCohortCsv, downloadCsv} from '@/api/cohort'
+import {deleteCohort} from '@/api/cohort'
+import {downloadCohortCsv, downloadCsv} from '@/api/cohort-csv'
 import {getCsvExportColumns, getCsvExportColumnsSelected} from '@/lib/berkeley-utils'
 import {useCohortStore} from '@/stores/cohort-edit-session'
