+++ src/views/AdmitStudents.vue
@@ -105,7 +105,7 @@ import Pagination from '@/components/util/Pagination'
 import SectionSpinner from '@/components/util/SectionSpinner'
 import SortBy from '@/components/student/SortBy'
 import {alertScreenReader, pluralize, putFocusNextTick, toInt} from '@/lib/utils'
-import {downloadCsv} from '@/api/cohort'
+import {downloadCsv} from '@/api/cohort-csv'
 import {getAdmitCsvExportColumns} from '@/lib/berkeley-utils'
 import {getAllAdmits} from '@/api/admit'
 import {useContextStore} from '@/stores/context'
