+++ tests/test_api/test_cohort_controller.py
@@ -24,13 +24,10 @@ ENHANCEMENTS, OR MODIFICATIONS.
 """
 
 from boac import std_commit
-from boac.api.csv_file_download_utils import get_students_csv_header_labels
-from boac.merged.sis_terms import current_term_id
 from boac.models.alert import Alert
 from boac.models.authorized_user import AuthorizedUser
 from boac.models.cohort_filter import CohortFilter
 from boac.models.curated_group import CuratedGroup
-import pytest
 import simplejson as json
 from tests.test_api.api_test_utils import all_cohorts_owned_by, api_cohort_create, api_cohort_events, \
     api_curated_group_add_students, api_curated_group_remove_student, api_get_cohort
@@ -1349,723 +1346,6 @@ class TestCohortPerFilters:
         )
         sids = sorted([s['sid'] for s in api_json['students']])
         assert sids == ['2345678901', '5678901234']
-
-
-class TestDownloadCohortCsv:
-
-    @classmethod
-    def setup_class(cls):
-        cls.coe_owned_cohort = next((c for c in all_cohorts_owned_by(coe_advisor_uid) if c['name'] == 'Radioactive Women and Men'), None)
-        cls.asc_owned_cohort = next((c for c in all_cohorts_owned_by(asc_advisor_uid) if c['name'] == 'All sports'), None)
-
-    @classmethod
-    def _api_download_cohort_csv(cls, client, cohort_id, csv_columns_selected, expected_status_code=200):
-        response = client.post(
-            '/api/cohort/download_csv',
-            data=json.dumps({
-                'cohortId': cohort_id,
-                'csvColumnsSelected': csv_columns_selected,
-            }),
-            content_type='application/json',
-        )
-        assert response.status_code == expected_status_code
-        return response.data
-
-    def test_download_csv_not_authenticated(self, client):
-        """API requires authentication."""
-        self._api_download_cohort_csv(
-            client,
-            self.coe_owned_cohort['id'],
-            csv_columns_selected=['sid'],
-            expected_status_code=401,
-        )
-
-    def test_download_csv_unauthorized(self, client, fake_auth):
-        """Sensitive COE attributes are not available to ASC advisor, when downloading CSV."""
-        fake_auth.login(asc_advisor_uid)
-        response = self._api_download_cohort_csv(
-            client,
-            cohort_id=self.asc_owned_cohort['id'],
-            csv_columns_selected=['coe_status'],
-        )
-        # If you strip away line breaks and double-quotes then you are left with nothing more than a column header.
-        data = response.decode('utf-8').replace('\n', '').replace('\r', '').replace('"', '')
-        assert data == 'CoE status'
-
-    def test_download_csv(self, client, fake_auth):
-        """Advisor can download cohort CSV."""
-        fake_auth.login(asc_advisor_uid)
-        expected_sids = ['11667051', '2345678901', '3456789012', '5678901234', '7890123456', '9100000000']
-        cohort = CohortFilter.create(
-            uid=asc_advisor_uid,
-            name='Download Me',
-            filter_criteria={
-                'cohortOwnerAcademicPlans': ['*'],
-            },
-        )
-        api_json = api_get_cohort(client, cohort_id=cohort['id'], include_students=True)
-        sids = sorted([student['sid'] for student in api_json['students']])
-        assert sids == expected_sids
-        data = self._api_download_cohort_csv(client, cohort['id'], csv_columns_selected=['sid'])
-        sids_in_csv = [s for s in data.decode('utf-8').split() if s.isdigit()]
-        assert sids_in_csv == expected_sids
-
-        # Another ASC advisor downloads same CSV
-        client.get('/api/auth/logout')
-        fake_auth.login('6446')
-        data = self._api_download_cohort_csv(client, cohort['id'], csv_columns_selected=['sid'])
-        sids_in_csv = [s for s in data.decode('utf-8').split() if s.isdigit()]
-        assert sids_in_csv == expected_sids
-
-
-class TestDownloadCsvPerFilters:
-
-    @classmethod
-    def _api_download_csv_per_filters(cls, client, json_data=(), expected_status_code=200):
-        response = client.post(
-            '/api/cohort/download_csv_per_filters',
-            data=json.dumps(json_data),
-            content_type='application/json',
-        )
-        assert response.status_code == expected_status_code
-        return response.json
-
-    def test_download_csv_not_authenticated(self, client):
-        """API requires authentication."""
-        self._api_download_csv_per_filters(client, expected_status_code=401)
-
-    def test_download_csv_with_empty(self, client, fake_auth):
-        """API requires non-empty input."""
-        fake_auth.login(coe_advisor_uid)
-        self._api_download_csv_per_filters(client, {'filters': ()}, expected_status_code=400)
-
-    def test_download_csv_unauthorized(self, client, fake_auth):
-        """ASC advisor is not allowed to query with COE attributes."""
-        fake_auth.login(asc_advisor_uid)
-        self._api_download_csv_per_filters(
-            client,
-            {
-                'filters': [
-                    {'key': 'coeAcademicStandings', 'value': ['P']},
-                ],
-                'csvColumnsSelected': [
-                    'first_name',
-                    'last_name',
-                    'sid',
-                ],
-            },
-            expected_status_code=403,
-        )
-
-    def test_download_csv(self, client, fake_auth):
-        """Advisor can download CSV with ALL students of cohort."""
-        fake_auth.login(coe_advisor_uid)
-        data = {
-            'filters': [
-                {'key': 'coeEthnicities', 'value': ['H', 'B']},
-            ],
-            'csvColumnsSelected': [
-                'first_name',
-                'last_name',
-                'sid',
-                'email',
-                'phone',
-                'college',
-                'majors',
-                'level_by_units',
-                'terms_in_attendance',
-                'expected_graduation_term',
-                'units_completed',
-                'term_gpa_2172',
-                'term_gpa_2175',
-                'cumulative_gpa',
-                'program_status',
-                'college_advisor',
-                'coe_status',
-                'course_activity',
-            ],
-        }
-        response = client.post(
-            '/api/cohort/download_csv_per_filters',
-            data=json.dumps(data),
-            content_type='application/json',
-        )
-        assert response.status_code == 200
-        assert 'csv' in response.content_type
-        csv = response.data.decode('UTF-8').split('\n')
-        # Verify that 'course_activity' related columns are present.
-        header_label_lookup = get_students_csv_header_labels(current_term_id())
-        expected_headers = ['first_name', 'last_name', 'sid', 'email', 'phone', 'majors', 'college', 'level_by_units',
-                            'terms_in_attendance', 'expected_graduation_term', 'units_completed', 'term_gpa_2172',
-                            'term_gpa_2175', 'cumulative_gpa', 'program_status', 'college_advisor', 'coe_status',
-                            'Class Name', 'Units', 'Mid-point Grade', 'Final Grade']
-        for expected_header in expected_headers:
-            expected_label = header_label_lookup.get(expected_header, expected_header)
-            assert expected_label in csv[0]
-        for row in csv:
-            if row.startswith('Deborah,Davies'):
-                assert '11667051' in row
-                assert 'Junior' in row
-                assert 'English BA' in row
-                assert ',Engineering; Undergrad Letters & Science,' in row
-                assert 'BURMESE 1A' in row or 'MED ST 205' in row or 'NUC ENG 124' in row or 'PHYSED 11' in row or 'SOCIOL 198' in row
-
-    def test_download_csv_custom_columns(self, client, fake_auth):
-        """Advisor can generate a CSV with the columns they want."""
-        fake_auth.login(coe_advisor_uid)
-        data = {
-            'filters': [
-                {'key': 'levels', 'value': 'Junior'},
-            ],
-            'csvColumnsSelected': [
-                'majors',
-                'college',
-                'level_by_units',
-                'terms_in_attendance',
-                'expected_graduation_term',
-                'units_completed',
-                'term_gpa_2172',
-                'cumulative_gpa',
-                'program_status',
-                'intended_majors',
-                'minors',
-            ],
-        }
-        response = client.post(
-            '/api/cohort/download_csv_per_filters',
-            data=json.dumps(data),
-            content_type='application/json',
-        )
-        assert response.status_code == 200
-        assert 'csv' in response.content_type
-        csv = response.data.decode('UTF-8').split('\n')
-        header_label_lookup = get_students_csv_header_labels(current_term_id())
-        expected_headers = ['majors', 'college', 'level_by_units', 'terms_in_attendance', 'expected_graduation_term',
-                            'units_completed', 'term_gpa_2172', 'cumulative_gpa', 'program_status', 'intended_majors',
-                            'minors']
-        for expected_header in expected_headers:
-            expected_label = header_label_lookup.get(expected_header, expected_header)
-            assert expected_label in csv[0]
-        for row in csv[1:]:
-            if row.startswith('Chemistry BS'):
-                assert ',Chemistry,' in row
-                assert 'Junior,4,Fall 2019,34,3.500,3.495,Active,' in row
-            elif row.startswith('English BA; Political Economy BA'):
-                assert ',Undergrad Letters & Science,' in row
-                assert 'Junior,5,Fall 2019,70,,3.005,Active,' in row
-            elif row:
-                pytest.fail(f'Unexpected CSV content: {row}')
-
-    admit_keys = [
-        'applyuc_cpid',
-        'cs_empl_id',
-        'residency_category',
-        'freshman_or_transfer',
-        'admit_term',
-        'admit_status',
-        'current_sir',
-        'college',
-        'first_name',
-        'middle_name',
-        'last_name',
-        'birthdate',
-        'daytime_phone',
-        'mobile',
-        'permanent_street_1',
-        'permanent_street_2',
-        'permanent_city',
-        'permanent_region',
-        'permanent_postal',
-        'permanent_country',
-        'xethnic',
-        'hispanic',
-        'urem',
-        'first_generation_college',
-        'parent_1_education_level',
-        'parent_2_education_level',
-        'highest_parent_education_level',
-        'hs_unweighted_gpa',
-        'hs_weighted_gpa',
-        'transfer_gpa',
-        'application_fee_waiver_flag',
-        'foster_care_flag',
-        'family_is_single_parent',
-        'student_is_single_parent',
-        'family_dependents_num',
-        'student_dependents_num',
-        'family_income',
-        'student_income',
-        'is_military_dependent',
-        'military_status',
-        'reentry_status',
-        'athlete_status',
-        'summer_bridge_status',
-        'last_school_lcff_plus_flag',
-        'special_program_cep',
-        'us_citizenship_status',
-        'us_non_citizen_status',
-        'citizenship_country',
-        'permanent_residence_country',
-        'non_immigrant_visa_current',
-        'non_immigrant_visa_planned',
-        'uid',
-    ]
-
-    @classmethod
-    def _api_download_admit_csv(cls, client, expected_status_code=200):
-        data = {
-            'filters': [
-                {'key': 'isFirstGenerationCollege', 'value': True},
-            ],
-            'csvColumnsSelected': cls.admit_keys,
-            'domain': 'admitted_students',
-        }
-        response = client.post(
-            '/api/cohort/download_csv_per_filters',
-            data=json.dumps(data),
-            content_type='application/json',
-        )
-        assert response.status_code == expected_status_code
-        return response
-
-    def test_download_csv_admit_domain(self, client, fake_auth):
-        """Download CSV admit domain."""
-        fake_auth.login(ce3_advisor_uid)
-        response = self._api_download_admit_csv(client)
-        assert 'csv' in response.content_type
-        csv = str(response.data)
-        assert ','.join(self.admit_keys) in csv
-        assert (
-            '19938035,00005852,RES,Transfer,Spring,No,No,College of Letters and Science,'
-            'Ralph,,Burgess,1984-09-04,984.110.7693x347,681-857-8070,9590 Chang Extensions,'
-            'Suite 478,East Jacobton,NY,55531,United States,International,F,No,Yes,MasterDegree,'
-            '3 - High School Graduate,,0.86,0.51,2.47,FeeWaiver,Y,,,05,02,41852,942,Y,'
-            'ReserveOfficersTrainingProgram,No,,,,,Citizen,,United States,,,,123'
-        ) in csv
-        assert (
-            '98002344,00029117,INT,Freshman,Spring,No,No,College of Engineering,Daniel,J,Mcknight,1993-07-06,859-319-8215x8689,'
-            '231.865.8093,87758 Brown Throughway,Suite 657,West Andrea,M,25101,United States,White,T,,'
-            'Yes,,5 - College Attended,,2.51,2.7,3.23,,,,Y,0,02,23915,426,Y,,,Committed,,1,'
-            'Destination College,Citizen,,United States,,,,'
-        ) in csv
-
-    def test_admit_domain_denies_non_ce3_advisor(self, user_factory, client, fake_auth):
-        advisor = user_factory(dept_codes=['GUEST'])
-        fake_auth.login(advisor.uid)
-        self._api_download_admit_csv(client, 404)
-
-
-class TestCohortFilterOptions:
-
-    @classmethod
-    def _api_cohort_filter_options(cls, client, expected_status_code=200, json_data=(), owner_uid=None):
-        response = client.post(
-            f'/api/cohort/filter_options/{owner_uid}',
-            data=json.dumps(json_data),
-            content_type='application/json',
-        )
-        assert response.status_code == expected_status_code
-        return response.json
-
-    def test_filter_options_api_not_authenticated(self, client):
-        """Menu API cohort-filter-options requires authentication."""
-        self._api_cohort_filter_options(client, expected_status_code=401)
-
-    def test_filter_options_with_nothing_disabled(self, client, fake_auth):
-        """Menu API with all menu options available."""
-        fake_auth.login(coe_advisor_uid)
-        api_json = self._api_cohort_filter_options(
-            client,
-            json_data={'existingFilters': []},
-            owner_uid=coe_advisor_uid,
-        )
-        for label, option_group in api_json.items():
-            for entry in option_group:
-                assert 'disabled' not in entry
-                if entry['type']['ux'] == 'dropdown':
-                    for option in entry['options']:
-                        assert 'disabled' not in option
-
-    def test_filter_options_for_guest_user(self, user_factory, client, fake_auth):
-        """Filter options available to GUEST user."""
-        advisor = user_factory(dept_codes=['GUEST'])
-        fake_auth.login(advisor.uid)
-        api_json = self._api_cohort_filter_options(
-            client,
-            json_data={'existingFilters': []},
-            owner_uid=advisor.uid,
-        )
-        assert len(api_json)
-        assert 'options' in list(api_json.values())[0][0]
-
-    def test_filter_options_for_user_of_type_other(self, user_factory, client, fake_auth):
-        """Filter options available to ZZZZZ user."""
-        advisor = user_factory(dept_codes=['ZZZZZ'])
-        fake_auth.login(advisor.uid)
-        api_json = self._api_cohort_filter_options(
-            client,
-            json_data={'existingFilters': []},
-            owner_uid=advisor.uid,
-        )
-        assert len(api_json)
-        assert 'options' in list(api_json.values())[0][0]
-
-    def test_filter_options_my_students_for_me(self, client, fake_auth):
-        """Returns user's own academic plans under 'My Students'."""
-        fake_auth.login(coe_advisor_uid)
-        api_json = self._api_cohort_filter_options(
-            client,
-            json_data={'existingFilters': []},
-            owner_uid=coe_advisor_uid,
-        )
-        my_students = next(opt for label, group in api_json.items() for opt in group if opt['label']['primary'] == 'My Students')
-        assert len(my_students['options']) == 5
-        assert {'name': 'All plans', 'value': '*'} in my_students['options']
-        assert {'name': 'Bioengineering BS', 'value': '16288U'} in my_students['options']
-        assert {'name': 'Engineering Undeclared UG', 'value': '162B0U'} in my_students['options']
-        assert {'name': 'BioE/MSE Joint Major BS', 'value': '162B3U'} in my_students['options']
-        assert {'name': 'Bioengineering UG', 'value': '16I010U'} in my_students['options']
-
-    def test_filter_options_my_students_for_not_me(self, client, fake_auth):
-        """Returns another user's academic plans under 'My Students'."""
-        fake_auth.login(coe_advisor_uid)
-        api_json = self._api_cohort_filter_options(
-            client,
-            json_data={'existingFilters': []},
-            owner_uid=asc_advisor_uid,
-        )
-        my_students = next(opt for label, group in api_json.items() for opt in group if opt['label']['primary'] == 'My Students')
-        assert len(my_students['options']) == 4
-        assert {'name': 'All plans', 'value': '*'} in my_students['options']
-        assert {'name': 'English BA', 'value': '25345U'} in my_students['options']
-        assert {'name': 'English UG', 'value': '25I039U'} in my_students['options']
-        assert {'name': 'Medieval Studies UG', 'value': '25I054U'} in my_students['options']
-
-    def test_filter_options_with_category_disabled(self, client, fake_auth):
-        """The transfer option is disabled if it is in existing-filters."""
-        fake_auth.login(coe_advisor_uid)
-        api_json = self._api_cohort_filter_options(
-            client,
-            json_data={'existingFilters': [{'key': 'transfer', 'value': True}]},
-            owner_uid=coe_advisor_uid,
-        )
-        assert len(api_json.keys())
-        for label, option_group in api_json.items():
-            for entry in option_group:
-                if entry['key'] == 'transfer':
-                    assert entry['disabled'] is True
-                else:
-                    assert 'disabled' not in entry
-
-    def test_filter_options_with_one_disabled(self, client, fake_auth):
-        """The 'Freshman' sub-menu option is disabled if it is already in cohort filter set."""
-        fake_auth.login(coe_advisor_uid)
-        api_json = self._api_cohort_filter_options(
-            client,
-            json_data={
-                'existingFilters':
-                    [
-                        {'key': 'levels', 'value': 'Freshman'},
-                        {'key': 'levels', 'value': 'Sophomore'},
-                        {'key': 'levels', 'value': 'Junior'},
-                        {'key': 'coeAdvisorLdapUids', 'value': '1022796'},
-                    ],
-            },
-            owner_uid=coe_advisor_uid,
-        )
-        assert len(api_json.keys())
-        assertion_count = 0
-        for label, opt_group in api_json.items():
-            for entry in opt_group:
-                # All top-level category menus are enabled
-                assert 'disabled' not in entry
-                if entry['key'] == 'levels':
-                    for option in entry['options']:
-                        disabled = option.get('disabled')
-                        if option['value'] in ['Freshman', 'Sophomore', 'Junior']:
-                            assert disabled is True
-                            assertion_count += 1
-                        else:
-                            assert disabled is None
-                else:
-                    assert 'disabled' not in entry
-        assert assertion_count == 3
-
-    def test_all_options_in_category_disabled(self, client, fake_auth):
-        """Disable the category if all its options are in existing-filters."""
-        fake_auth.login(coe_advisor_uid)
-        api_json = self._api_cohort_filter_options(
-            client,
-            json_data={
-                'existingFilters':
-                    [
-                        {'key': 'cohortOwnerAcademicPlans', 'value': '*'},
-                        {'key': 'levels', 'value': 'Doctoral Candidate > 6'},
-                        {'key': 'levels', 'value': 'Doctoral Candidate <= 6'},
-                        {'key': 'levels', 'value': 'Doctoral Pre-Candidacy'},
-                        {'key': 'levels', 'value': 'Masters/Professional'},
-                        {'key': 'levels', 'value': 'Senior'},
-                        {'key': 'levels', 'value': 'Junior'},
-                        {'key': 'levels', 'value': 'Sophomore'},
-                        {'key': 'levels', 'value': 'Freshman'},
-                        {'key': 'visaTypes', 'value': '*'},
-                    ],
-            },
-            owner_uid=coe_advisor_uid,
-        )
-        for label, option_group in api_json.items():
-            for entry in option_group:
-                if entry['key'] == 'cohortOwnerAcademicPlans':
-                    assert entry.get('disabled') is True
-                elif entry['key'] == 'levels':
-                    assert entry.get('disabled') is True
-                    for option in entry['options']:
-                        assert option.get('disabled') is True
-                elif entry['key'] == 'visaTypes':
-                    assert entry.get('disabled') is True
-                else:
-                    assert 'disabled' not in entry
-
-    def test_range_of_entering_terms(self, user_factory, client, fake_auth):
-        """Cohort filter: Range of entering terms."""
-        advisor = user_factory(dept_codes=['GUEST'])
-        fake_auth.login(advisor.uid)
-        api_json = self._api_cohort_filter_options(
-            client,
-            json_data={'existingFilters': []},
-            owner_uid=advisor.uid,
-        )
-        entering_terms_filter = next((f for f in api_json['Academic'] if f['key'] == 'enteringTerms'), None)
-        assert entering_terms_filter
-        filter_options = entering_terms_filter.get('options')
-        assert len(filter_options) == 4
-        assert [o['name'] for o in filter_options] == ['2015 Fall', '2015 Summer', '2015 Spring', '1993 Fall']
-
-    def test_range_of_expected_grad_terms(self, user_factory, client, fake_auth):
-        """Cohort filter: Range of expected grad terms."""
-        advisor = user_factory(dept_codes=['GUEST'])
-        fake_auth.login(advisor.uid)
-        api_json = self._api_cohort_filter_options(
-            client,
-            json_data={'existingFilters': []},
-            owner_uid=advisor.uid,
-        )
-        entering_terms_filter = next((f for f in api_json['Academic'] if f['key'] == 'expectedGradTerms'), None)
-        assert entering_terms_filter
-        filter_options = entering_terms_filter.get('options')
-        assert len(filter_options['Past']) == 1
-        assert filter_options['Past'][0]['name'] == '1997 Fall'
-
-    def test_range_of_majors(self, user_factory, client, fake_auth):
-        """Cohort filter: Range of major terms."""
-        advisor = user_factory(dept_codes=['GUEST'])
-        fake_auth.login(advisor.uid)
-        api_json = self._api_cohort_filter_options(
-            client,
-            json_data={'existingFilters': []},
-            owner_uid=advisor.uid,
-        )
-        majors_filter = next((f for f in api_json['Academic'] if f['key'] == 'majors'), None)
-        assert {'name': 'Chemistry BS', 'value': 'Chemistry BS'} in majors_filter['options']
-        assert {'name': 'Nuclear Engineering BS', 'value': 'Nuclear Engineering BS'} in majors_filter['options']
-        assert {'name': 'Mathematics PhD', 'value': 'Mathematics PhD'} not in majors_filter['options']
-
-    def test_range_of_graduate_programs(self, user_factory, client, fake_auth):
-        """Cohort filter: Range of graduate programs."""
-        advisor = user_factory(dept_codes=['GUEST'])
-        fake_auth.login(advisor.uid)
-        api_json = self._api_cohort_filter_options(
-            client,
-            json_data={'existingFilters': []},
-            owner_uid=advisor.uid,
-        )
-        majors_filter = next((f for f in api_json['Academic'] if f['key'] == 'graduatePrograms'), None)
-        assert len(majors_filter['options']) == 1
-        assert majors_filter['options'][0] == {'name': 'Mathematics PhD', 'value': 'Mathematics PhD'}
-
-    def test_no_curated_group_options(self, client, fake_auth):
-        """User with no curated groups gets no cohort filter option where key='curatedGroupIds'."""
-        fake_auth.login(asc_and_coe_advisor_uid)
-        user_id = AuthorizedUser.get_id_per_uid(asc_and_coe_advisor_uid)
-        assert not CuratedGroup.get_curated_groups(user_id)
-        api_json = self._api_cohort_filter_options(
-            client,
-            json_data={'existingFilters': []},
-            owner_uid=asc_and_coe_advisor_uid,
-        )
-        verified = False
-        for label, option_group in api_json.items():
-            for filter_ in option_group:
-                if filter_['key'] == 'curatedGroupIds':
-                    assert filter_['disabled'] is True
-                    verified = True
-        assert verified
-
-    def test_invalid_domain_value(self, user_factory, client, fake_auth):
-        """Cohort filter: Invalid domain value."""
-        advisor = user_factory(dept_codes=['GUEST'])
-        fake_auth.login(advisor.uid)
-        self._api_cohort_filter_options(
-            client,
-            expected_status_code=400,
-            json_data={
-                'domain': 'this_is_an_invalid_domain',
-                'existingFilters': [],
-            },
-            owner_uid=advisor.uid,
-        )
-
-    def test_admitted_students_domain_denied(self, user_factory, client, fake_auth):
-        """Cohort filter: Denied access to Admitted students."""
-        advisor = user_factory(dept_codes=['GUEST'])
-        fake_auth.login(advisor.uid)
-        self._api_cohort_filter_options(
-            client,
-            expected_status_code=404,
-            json_data={
-                'domain': 'admitted_students',
-                'existingFilters': [],
-            },
-            owner_uid=advisor.uid,
-        )
-
-    def test_admitted_students_domain(self, client, fake_auth):
-        """Cohort filter: Admitted students."""
-        fake_auth.login(ce3_advisor_uid)
-        api_json = self._api_cohort_filter_options(
-            client,
-            json_data={
-                'domain': 'admitted_students',
-                'existingFilters': [],
-            },
-            owner_uid=ce3_advisor_uid,
-        )
-        assert len(api_json)
-        for label, option_group in api_json.items():
-            for entry in option_group:
-                # Verify the 'default' filters are not present.
-                assert entry['key'] != 'unitRanges'
-                assert entry['domain'] == 'admitted_students'
-
-
-class TestTranslateToFilterOptions:
-
-    @classmethod
-    def _api_translate_to_filter_options(cls, client, expected_status_code=200, json_data=(), owner_uid=None):
-        response = client.post(
-            f'/api/cohort/translate_to_filter_options/{owner_uid}',
-            data=json.dumps(json_data),
-            content_type='application/json',
-        )
-        assert response.status_code == expected_status_code
-        return response.json
-
-    def test_translate_criteria_when_empty(self, client, fake_auth):
-        """Empty criteria translates to zero rows."""
-        fake_auth.login(coe_advisor_uid)
-        assert self._api_translate_to_filter_options(
-            client,
-            json_data={
-                'criteria': {},
-            },
-            owner_uid=coe_advisor_uid,
-        ) == []
-
-    def test_translate_criteria_with_boolean(self, client, fake_auth):
-        """Filter-criteria with boolean is properly translated."""
-        fake_auth.login(coe_advisor_uid)
-        api_json = self._api_translate_to_filter_options(
-            client,
-            json_data={'criteria': {'isInactiveCoe': False}},
-            owner_uid=coe_advisor_uid,
-        )
-        assert len(api_json) == 1
-        assert api_json[0]['label']['primary'] == 'Inactive (COE)'
-        assert api_json[0]['key'] == 'isInactiveCoe'
-        assert api_json[0]['value'] is False
-
-    def test_translate_criteria_with_array(self, client, fake_auth):
-        """Filter-criteria with array is properly translated."""
-        fake_auth.login(coe_advisor_uid)
-        api_json = self._api_translate_to_filter_options(
-            client,
-            json_data={
-                'criteria': {
-                    'levels': ['Freshman', 'Sophomore'],
-                    'majors': ['Chemistry BS', 'Nuclear Engineering BS'],
-                },
-            },
-            owner_uid=coe_advisor_uid,
-        )
-        assert len(api_json) == 4
-        # Levels
-        assert api_json[0]['label']['primary'] == api_json[1]['label']['primary'] == 'Level'
-        assert api_json[0]['key'] == api_json[1]['key'] == 'levels'
-        assert api_json[0]['value'] == 'Freshman'
-        assert api_json[1]['value'] == 'Sophomore'
-        # Majors
-        assert api_json[2]['label']['primary'] == api_json[3]['label']['primary'] == 'Major'
-        assert api_json[2]['key'] == api_json[3]['key'] == 'majors'
-        assert api_json[2]['value'] == 'Chemistry BS'
-        assert api_json[3]['value'] == 'Nuclear Engineering BS'
-
-    def test_handle_last_name_ranges(self, client, fake_auth):
-        """Filter-criteria with last_name range is properly translated."""
-        fake_auth.login(coe_advisor_uid)
-        last_name_ranges = [
-            {'min': 'B', 'max': 'D'},
-            {'min': 'M', 'max': 'Z'},
-        ]
-        api_json = self._api_translate_to_filter_options(
-            client,
-            json_data={
-                'criteria': {
-                    'lastNameRanges': last_name_ranges,
-                },
-            },
-            owner_uid=coe_advisor_uid,
-        )
-        assert len(api_json) == 2
-
-        def _verify(index):
-            assert api_json[index]['label']['primary'] == 'Last Name'
-            assert api_json[index]['key'] == 'lastNameRanges'
-            assert api_json[index]['value'] == last_name_ranges[index]
-        _verify(0)
-        _verify(1)
-
-    def test_translate_criteria_my_students_for_me(self, client, fake_auth):
-        """User's own 'My Students' criteria are properly translated."""
-        fake_auth.login(coe_advisor_uid)
-        api_json = self._api_translate_to_filter_options(
-            client,
-            json_data={
-                'criteria': {
-                    'cohortOwnerAcademicPlans': ['*'],
-                },
-            },
-            owner_uid=coe_advisor_uid,
-        )
-        assert len(api_json) == 1
-        assert api_json[0]['label']['primary'] == 'My Students'
-        assert api_json[0]['key'] == 'cohortOwnerAcademicPlans'
-        assert api_json[0]['value'] == '*'
-
-    def test_translate_criteria_my_students_for_not_me(self, client, fake_auth):
-        """Another user's 'My Students' criteria are properly translated."""
-        fake_auth.login(coe_advisor_uid)
-        api_json = self._api_translate_to_filter_options(
-            client,
-            json_data={'criteria': {'cohortOwnerAcademicPlans': ['25I039U', '25I054U']}},
-            owner_uid=asc_advisor_uid,
-        )
-        assert len(api_json) == 2
-        assert api_json[0]['label']['primary'] == 'My Students'
-        assert api_json[0]['key'] == 'cohortOwnerAcademicPlans'
-        assert api_json[0]['value'] == '25I039U'
-        assert api_json[1]['label']['primary'] == 'My Students'
-        assert api_json[1]['key'] == 'cohortOwnerAcademicPlans'
-        assert api_json[1]['value'] == '25I054U'
 
 
 def _new_undeclared_cohort(client):
